split into actor and erlang-adapted modules
