+++ Cargo.toml
@@ -14,6 +14,14 @@ categories = ["concurrency"]
 [badges]
 maintenance = { status = "experimental" }
 
+[dev-dependencies]
+criterion = "0.3.3"
+
 [dependencies]
 async-task = "4.0.3"
 smol = "1.2.5"
+thiserror = "1.0.23"
+
+[[bench]]
+name = "general"
+harness = false
